fix tests

+++ src/tests/EntityGraphQL.Tests/LinqQuery/LinqCompilerTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using EntityGraphQL.Schema;
 using EntityGraphQL.Extensions;
 using EntityGraphQL.Compiler;
+using System;
 
 namespace EntityGraphQL.LinqQuery.Tests
 {
@@ -16,14 +17,14 @@ namespace EntityGraphQL.LinqQuery.Tests
         public void CompilesNumberConstant()
         {
             var exp = EqlCompiler.Compile("3");
-            Assert.Equal(3, exp.Execute());
+            Assert.Equal((UInt64)3, exp.Execute());
         }
 
         [Fact]
         public void CompilesNegitiveNumberConstant()
         {
             var exp = EqlCompiler.Compile("-43");
-            Assert.Equal(-43, exp.Execute());
+            Assert.Equal((Int64)(-43), exp.Execute());
         }
 
         [Fact]
@@ -121,7 +122,7 @@ namespace EntityGraphQL.LinqQuery.Tests
         {
             // tells it how to read it
             var exp = EqlCompiler.Compile("(someRelation.relation.id = 99) ? 100 : 66", SchemaBuilder.FromObject<TestSchema>());
-            Assert.Equal(100, exp.Execute(new TestSchema()));
+            Assert.Equal((UInt64)100, exp.Execute(new TestSchema()));
         }
 
         [Fact]
@@ -129,32 +130,32 @@ namespace EntityGraphQL.LinqQuery.Tests
         {
             // tells it how to read it
             var exp = EqlCompiler.Compile("(someRelation.relation.id = 98) ? 100 : 66", SchemaBuilder.FromObject<TestSchema>());
-            Assert.Equal(66, exp.Execute(new TestSchema()));
+            Assert.Equal((UInt64)66, exp.Execute(new TestSchema()));
         }
 
         [Fact]
         public void CompilesIfThenElseTrue()
         {
             var exp = EqlCompiler.Compile("if someRelation.relation.id = 99 then 100 else 66", SchemaBuilder.FromObject<TestSchema>());
-            Assert.Equal(100, exp.Execute(new TestSchema()));
+            Assert.Equal((UInt64)100, exp.Execute(new TestSchema()));
         }
         [Fact]
         public void CompilesIfThenElseFalse()
         {
             var exp = EqlCompiler.Compile("if someRelation.relation.id = 33 then 100 else 66", SchemaBuilder.FromObject<TestSchema>());
-            Assert.Equal(66, exp.Execute(new TestSchema()));
+            Assert.Equal((UInt64)66, exp.Execute(new TestSchema()));
         }
         [Fact]
         public void CompilesBinaryWithIntAndUint()
         {
             var exp = EqlCompiler.Compile("if someRelation.UnisgnedInt = 33 then 100 else 66", SchemaBuilder.FromObject<TestSchema>());
-            Assert.Equal(66, exp.Execute(new TestSchema()));
+            Assert.Equal((UInt64)66, exp.Execute(new TestSchema()));
         }
         [Fact]
         public void CompilesBinaryWithNullableAndNonNullable()
         {
             var exp = EqlCompiler.Compile("if someRelation.NullableInt = 8 then 100 else 66", SchemaBuilder.FromObject<TestSchema>());
-            Assert.Equal(100, exp.Execute(new TestSchema()));
+            Assert.Equal((UInt64)100, exp.Execute(new TestSchema()));
         }
         [Fact]
         public void CanUseCompiledExpressionInWhereMethod()
